Revert "Issue #237 renamed package DNAnalzer to analyzer."
This reverts commit 46d250119d5cac164a853eb585812e818d0ed1eb.
