+++ CITATION.cff
@@ -8,4 +8,4 @@ date-released: 2022-10-10
 message: "If you use this software, please cite it using these metadata."
 title: "A highly efficient, powerful, and feature-rich machine learning algorithm for analyzing DNA sequences"
 version: "2.0.0"
-url: "https://github.com/Verisimilitude11/analyzer"
+url: "https://github.com/Verisimilitude11/DNAnalyzer"
