+++ src/main/java/DNAnalyzer/BasePairCounter.java
