+++ src/main/java/DNAnalyzer/ReadingFrames.java
@@ -9,9 +9,9 @@
  * For further inquiries, please reach out to contact@dnanalyzer.live
  */
 
-package analyzer;
+package DNAnalyzer;
 
-import analyzer.codon.*;
+import DNAnalyzer.codon.*;
 
 import java.util.HashMap;
 import java.util.Map;
