+++ src/main/java/DNAnalyzer/codon/CodonDataConstants.java
@@ -8,11 +8,11 @@
  *
  * For further inquiries, please reach out to contact@dnanalyzer.live
  */
-package analyzer.codon;
+package DNAnalyzer.codon;
 
 import static java.util.Map.entry;
 
-import analyzer.aminoAcid.*;
+import DNAnalyzer.aminoAcid.*;
 
 import java.util.*;
