+++ src/main/java/DNAnalyzer/protein/ProteinAnalysis.java
@@ -9,11 +9,11 @@
  * For further inquiries, please reach out to contact@dnanalyzer.live
  */
 
-package analyzer.protein;
+package DNAnalyzer.protein;
 
 import java.util.List;
 
-import analyzer.Properties;
+import DNAnalyzer.Properties;
 
 /**
  * Prints properties of the proteins in the DNA.
