+++ src/main/java/DNAnalyzer/protein/ProteinFinder.java
