+++ src/test/java/DNAnalyzer/MainTest.java
@@ -1,11 +1,11 @@
 /*
  * This Java source file was generated by the Gradle 'init' task.
  */
-package analyzer;
+package DNAnalyzer;
 
 public class MainTest {
     // @Test
     public void mainClassshouldExist() throws ClassNotFoundException {
-        Class.forName("analyzer.Main");
+        Class.forName("DNAnalyzer.Main");
     }
 }
